Revert "fix(explorer): fetch address balances/metadata via server fn (avoid SSR self-fetch)"
This reverts commit 44de72224ce57eecd4fe2720c46ac7782ab36be3.
