+++ apps/explorer/src/routes/_layout/address/$address.tsx
@@ -90,16 +90,13 @@ import {
 	holdersQueryOptions,
 	transfersQueryOptions,
 } from '#lib/queries/tokens'
-
+import { getApiUrl } from '#lib/env.ts'
 import { areUsdPricedTokens } from '#lib/pricing'
 import {
 	fetchAddressBalancesData,
 	MAX_TOKENS,
 } from '#lib/server/address-balances'
-import {
-	buildAddressTxMetadata,
-	fetchAddressMetadata,
-} from '#lib/server/address-metadata'
+import { buildAddressTxMetadata } from '#lib/server/address-metadata'
 import {
 	fetchAddressOldestTx,
 	fetchAddressTxStats,
@@ -747,6 +744,14 @@ function RouteComponent() {
 	)
 }
 
+async function fetchAddressMetadata(address: Address.Address) {
+	const response = await fetch(getApiUrl(`/api/address/metadata/${address}`), {
+		headers: { 'Content-Type': 'application/json' },
+	})
+	if (!response.ok) throw new Error('Failed to fetch address metadata')
+	return response.json() as Promise<AddressOgMetadata>
+}
+
 async function fetchAddressOgMetadata(
 	address: Address.Address,
 ): Promise<AddressOgMetadata> {
@@ -784,7 +789,7 @@ async function fetchAddressOgBalances(address: Address.Address) {
 function addressMetadataQueryOptions(address: Address.Address) {
 	return {
 		queryKey: ['address-metadata', address] as const,
-		queryFn: () => fetchAddressMetadata({ data: { address } }),
+		queryFn: () => fetchAddressMetadata(address),
 		staleTime: 30_000,
 	}
 }
